(NPUP-15) Port relationship operators to the dispatcher.
Porting the relationship operators' implementation to the dispatcher.
