+++ lib/include/puppet/compiler/evaluation/operators/binary/out_edge_subscribe.hpp
@@ -1,0 +1,23 @@
+/**
+ * @file
+ * Declares the out-edge-subscribe operator.
+ */
+#pragma once
+
+#include "descriptor.hpp"
+
+namespace puppet { namespace compiler { namespace evaluation { namespace operators { namespace binary {
+
+    /**
+     * Implements the out-edge-subscribe (~>) operator.
+     */
+    struct out_edge_subscribe
+    {
+        /**
+         * Create a binary operator descriptor.
+         * @return Returns the binary operator descriptor representing this operator.
+         */
+        static descriptor create_descriptor();
+    };
+
+}}}}}  // namespace puppet::compiler::evaluation::operators::binary
