+++ lib/src/compiler/evaluation/evaluator.cc
@@ -6,7 +6,6 @@
 #include <puppet/compiler/evaluation/functions/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/binary/call_context.hpp>
 #include <puppet/compiler/evaluation/operators/unary/call_context.hpp>
-#include <puppet/compiler/evaluation/operators/relationship.hpp>
 #include <puppet/compiler/evaluation/operators/right_shift.hpp>
 #include <puppet/compiler/evaluation/operators/splat.hpp>
 #include <puppet/compiler/exceptions.hpp>
@@ -911,10 +910,6 @@ namespace puppet { namespace compiler { namespace evaluation {
         ast::binary_operation const& operation)
     {
         static const unordered_map<binary_operator, function<value(operators::binary_operator_context const&)>, boost::hash<binary_operator>> binary_operators = {
-            { ast::binary_operator::in_edge,            operators::in_edge() },
-            { ast::binary_operator::in_edge_subscribe,  operators::in_edge_subscribe() },
-            { ast::binary_operator::out_edge,           operators::out_edge() },
-            { ast::binary_operator::out_edge_subscribe, operators::out_edge_subscribe() },
             { ast::binary_operator::right_shift,        operators::right_shift() }
         };
